pyproject.toml
[project]
name = "kryten-economy"
version = "0.1.0"
description = "Channel engagement currency microservice for CyTube channels"
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
license = {text = "MIT"}
authors = [
    {name = "grobertson"},
]
keywords = [
    "cytube",
    "economy",
    "currency",
    "chat",
    "bot",
    "nats",
    "microservice",
    "kryten",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Typing :: Typed",
    "Framework :: AsyncIO",
]
dependencies = [
    "kryten-py>=0.11.5",
    "pyyaml>=6.0,<7.0.0",
    "pydantic>=2.0,<3.0.0",
    "aiohttp>=3.9.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<0.25.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.4.0,<1.0.0",
    "mypy>=1.10.0,<2.0.0",
    "build>=1.0.0",
    "twine>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/grobertson/kryten-economy"
Repository = "https://github.com/grobertson/kryten-economy"
Issues = "https://github.com/grobertson/kryten-economy/issues"

[project.scripts]
kryten-economy = "kryten_economy.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kryten_economy"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
